+++ packages/ricos-styles/web/src/document-style/text-style.ts
@@ -2,11 +2,15 @@ import type { TextStyle as TextStyleRichContent } from 'ricos-schema';
 import type { CustomTextualStyle } from 'ricos-types';
 import type { TextStyle } from '../models/text-style';
 
-export default class RicosTextStyle implements TextStyle {
+export class RicosTextStyle implements TextStyle {
   textStyle: Omit<TextStyleRichContent, 'textAlignment'>;
 
-  constructor(textStyle: Omit<TextStyleRichContent, 'textAlignment'>) {
+  private constructor(textStyle: Omit<TextStyleRichContent, 'textAlignment'>) {
     this.textStyle = textStyle;
+  }
+
+  static of(textStyle: Omit<TextStyleRichContent, 'textAlignment'>): RicosTextStyle {
+    return new RicosTextStyle(textStyle || {});
   }
 
   getTextStyle: TextStyle['getTextStyle'] = () => this.textStyle;
@@ -14,7 +18,7 @@ export default class RicosTextStyle implements TextStyle {
   static fromCustomStyle = (customStyle: CustomTextualStyle): TextStyle => {
     const { lineHeight } = customStyle;
     const textStyle = { lineHeight } as TextStyleRichContent;
-    return new RicosTextStyle(textStyle);
+    return RicosTextStyle.of(textStyle);
   };
 
   toCustomStyle: TextStyle['toCustomStyle'] = () => {
@@ -22,6 +26,6 @@ export default class RicosTextStyle implements TextStyle {
   };
 
   overrideWith: TextStyle['overrideWith'] = textStyle => {
-    return new RicosTextStyle({ ...this.textStyle, ...textStyle });
+    return RicosTextStyle.of({ ...this.textStyle, ...textStyle });
   };
 }